Move flags to a separate module.
- Add type to the flag interface, so that mistakes with flag names or types are checked.
